Add overlapping patches experience

+++ dl_inv_prob/dl.py
@@ -163,7 +163,7 @@ class DictionaryLearning(nn.Module):
         """
         out = torch.zeros(
             (self.n_matrices, self.dim_x, y.shape[2]),
-            dtype=torch.float,
+            dtype=torch.float64,
             device=self.device
         )
 
@@ -282,6 +282,7 @@ class DictionaryLearning(nn.Module):
 
         # Avoiding numerical instabitility in the step size
         future_step = min(10*t, 1e4)
+
         # future_step = t
         return future_step, end
 
@@ -364,6 +365,7 @@ class DictionaryLearning(nn.Module):
         loss : float
             Final value of the loss after training.
         """
+
         # Dimension
         self.dim_y = Y.shape[1]
 
@@ -371,21 +373,21 @@ class DictionaryLearning(nn.Module):
         if A is None:
             self.dim_signal = self.dim_y
             self.operator = torch.eye(
-                self.dim_y, device=self.device, dtype=torch.float
+                self.dim_y, device=self.device, dtype=torch.float64
             )[None, :]
         else:
             self.dim_signal = A.shape[2]
-            self.operator = torch.from_numpy(A).float().to(self.device)
+            self.operator = torch.from_numpy(A).double().to(self.device)
 
         self.n_matrices = self.operator.shape[0]
 
         # Covariance
         if cov_inv is None:
             self.cov_inv = torch.eye(
-                self.dim_signal, device=self.device, dtype=torch.float
+                self.dim_signal, device=self.device, dtype=torch.float64
             )
         else:
-            self.cov_inv = torch.from_numpy(cov_inv).float().to(self.device)
+            self.cov_inv = torch.from_numpy(cov_inv).double().to(self.device)
 
         # Dictionary
         if self.init_D is None:
@@ -398,10 +400,11 @@ class DictionaryLearning(nn.Module):
             if A is None:
                 dico = Y[0, :, choice]
             self.D = nn.Parameter(
-                torch.tensor(dico, device=self.device, dtype=torch.float)
+                torch.tensor(dico, device=self.device, dtype=torch.float64)
             )
         else:
-            dico_tensor = torch.from_numpy(self.init_D).float().to(self.device)
+            dico_tensor = torch.from_numpy(
+                self.init_D).double().to(self.device)
             self.D = nn.Parameter(dico_tensor)
 
         # Scaling and computing lipschitz
@@ -409,7 +412,7 @@ class DictionaryLearning(nn.Module):
         self.compute_lipschitz()
 
         # Data
-        self.Y_tensor = torch.from_numpy(Y).float().to(self.device)
+        self.Y_tensor = torch.from_numpy(Y).double().to(self.device)
 
         # Training
         loss = self.training_process()
@@ -429,7 +432,7 @@ class DictionaryLearning(nn.Module):
             Reconstructed signal.
         """
         with torch.no_grad():
-            y = torch.tensor(y, device=self.device, dtype=torch.float)
+            y = torch.tensor(y, device=self.device, dtype=torch.float64)
             x = self.forward(y)
             rec = torch.matmul(self.D, x)
 
@@ -484,7 +487,7 @@ class Inpainting(DictionaryLearning):
         """
         out = torch.zeros(
             (self.n_matrices, self.dim_x, y.shape[2]),
-            dtype=torch.float,
+            dtype=torch.float64,
             device=self.device
         )
 
@@ -542,16 +545,16 @@ class Inpainting(DictionaryLearning):
         self.dim_signal = Y.shape[1]
 
         # Mask
-        self.mask = torch.from_numpy(mask).float().to(self.device)
+        self.mask = torch.from_numpy(mask).double().to(self.device)
         self.n_matrices = self.mask.shape[0]
 
         # Covariance
         if cov_inv is None:
             self.cov_inv = torch.eye(
-                self.dim_signal, device=self.device, dtype=torch.float
+                self.dim_signal, device=self.device, dtype=torch.float64
             )
         else:
-            self.cov_inv = torch.from_numpy(cov_inv).float().to(self.device)
+            self.cov_inv = torch.from_numpy(cov_inv).double().to(self.device)
 
         # Dictionary
         if self.init_D is None:
@@ -561,10 +564,11 @@ class Inpainting(DictionaryLearning):
                 size=(self.dim_signal, self.n_components)
             )
             self.D = nn.Parameter(
-                torch.tensor(dico, device=self.device, dtype=torch.float)
+                torch.tensor(dico, device=self.device, dtype=torch.float64)
             )
         else:
-            dico_tensor = torch.from_numpy(self.init_D).float().to(self.device)
+            dico_tensor = torch.from_numpy(
+                self.init_D).double().to(self.device)
             self.D = nn.Parameter(dico_tensor)
 
         # Scaling and computing lipschitz
@@ -572,7 +576,7 @@ class Inpainting(DictionaryLearning):
         self.compute_lipschitz()
 
         # Data
-        self.Y_tensor = torch.from_numpy(Y).float().to(self.device)
+        self.Y_tensor = torch.from_numpy(Y).double().to(self.device)
 
         # Training
         loss = self.training_process()
